requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
pydantic==2.10.4
python-dotenv==1.0.1

# Data pipeline
xarray==2024.11.0
zarr==2.18.4
netCDF4==1.7.2
cdsapi==0.7.5
numpy==1.26.4
pandas==2.2.3
scipy==1.14.1
dask[complete]==2024.12.1

# ML Core
torch==2.5.1
scikit-learn==1.6.0
shap==0.46.0

# Forecasting
pytorch-forecasting==1.2.0
pytorch-lightning==2.4.0
mapie==0.9.1

# Explainability + LLM
google-generativeai==0.8.5

# Caching
redis==5.2.1

# GeoJSON
geojson==3.2.0
shapely==2.0.6

# Utilities
loguru==0.7.3
httpx==0.28.1
tqdm==4.67.1
